--- a/tensorflow/core/data/root_dataset.cc
+++ b/tensorflow/core/data/root_dataset.cc
@@ -62,7 +62,7 @@ Status RootDataset::FromOptions(DatasetBase* input, DatasetBase** output) {
     // check for deprecated autotune options in optimization_options
     // the condition below checks for the presence of the deprecated option
     // and decides whether to go with the deprecated options or not.
-    if(optimization_options.optional_autotune_buffers_case() ==
+    if(options.optimization_options().optional_autotune_buffers_case() ==
       OptimizationOptions::kAutotuneBuffers){
       if (options.optimization_options().autotune_buffers()) {
         params.autotune_algorithm = model::AutotuneAlgorithm::GRADIENT_DESCENT;