--- a/tensorflow/core/framework/dataset_options.proto
+++ b/tensorflow/core/framework/dataset_options.proto
@@ -127,6 +127,30 @@ enum ExternalStatePolicy {
   POLICY_FAIL = 2;
 }
 
+message AutotuneOptions {
+  // Whether to automatically tune performance knobs.
+  oneof optional_enabled {
+    bool enabled = 1;
+  }
+  // When autotuning is enabled (through autotune), determines whether to also
+  // autotune buffer sizes for datasets with parallelism.
+  oneof optional_autotune_buffers {
+    bool autotune_buffers = 2;
+  }
+  // When autotuning is enabled (through autotune), determines the CPU budget to
+  // use. Values greater than the number of schedulable CPU cores are allowed
+  // but may result in CPU contention.
+  oneof optional_cpu_budget {
+    int32 cpu_budget = 3;
+  }
+  // When autotuning is enabled (through autotune), determines the RAM budget to
+  // use. Values greater than the available RAM in bytes may result in OOM. If
+  // 0, defaults to half of the available RAM in bytes.
+  oneof optional_ram_budget {
+    int64 ram_budget = 4;
+  }
+}
+
 // Message stored with Dataset objects to control how datasets are processed and
 // optimized.
 message Options {
@@ -155,4 +179,6 @@ message Options {
   oneof optional_external_state_policy {
     ExternalStatePolicy external_state_policy = 6;
   }
+  // The autotune options associated with the dataset
+  AutotuneOptions autotune_options = 7;
 }