--- a/tensorflow/core/kernels/data/finalize_dataset_op.cc
+++ b/tensorflow/core/kernels/data/finalize_dataset_op.cc
@@ -38,11 +38,11 @@ void GetModelDatasetParams(const Options& options,
                            model::AutotuneAlgorithm* algorithm,
                            bool* cpu_budget, bool* ram_budget) {
   *algorithm = model::AutotuneAlgorithm::HILL_CLIMB;
-  if (options.optimization_options().autotune_buffers()) {
+  if (options.autotune_options().autotune_buffers()) {
     *algorithm = model::AutotuneAlgorithm::GRADIENT_DESCENT;
   }
-  *cpu_budget = options.optimization_options().autotune_cpu_budget();
-  *ram_budget = options.optimization_options().autotune_ram_budget();
+  *cpu_budget = options.autotune_options().cpu_budget();
+  *ram_budget = options.autotune_options().ram_budget();
 }
 
 void MakeDatasetHelper(OpKernelContext* ctx, bool has_captured_ref,