--- a/tensorflow/python/data/experimental/ops/autotune_options.py
+++ b/tensorflow/python/data/experimental/ops/autotune_options.py
@@ -1,4 +1,4 @@
-# Copyright 2018 The TensorFlow Authors. All Rights Reserved.
+# Copyright 2021 The TensorFlow Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.