--- a/tensorflow/python/data/kernel_tests/BUILD
+++ b/tensorflow/python/data/kernel_tests/BUILD
@@ -645,6 +645,7 @@ tf_py_test(
     deps = [
         ":test_base",
         "//tensorflow/python:client_testlib",
+        "//tensorflow/python/data/experimental/ops:autotune_options",
         "//tensorflow/python/data/experimental/ops:optimization_options",
         "//tensorflow/python/data/experimental/ops:testing",
         "//tensorflow/python/data/experimental/ops:threading_options",