--- a/tensorflow/tools/api/golden/v1/tensorflow.data.-autotune-options.pbtxt
+++ b/tensorflow/tools/api/golden/v1/tensorflow.data.-autotune-options.pbtxt
@@ -1,0 +1,26 @@
+path: "tensorflow.data.AutotuneOptions"
+tf_class {
+  is_instance: "<class \'tensorflow.python.data.experimental.ops.autotune_options.AutotuneOptions\'>"
+  is_instance: "<class \'tensorflow.python.data.util.options.OptionsBase\'>"
+  is_instance: "<type \'object\'>"
+  member {
+    name: "cpu_budget"
+    mtype: "<type \'property\'>"
+  }
+  member {
+    name: "enabled"
+    mtype: "<type \'property\'>"
+  }
+  member {
+    name: "experimental_autotune_buffers"
+    mtype: "<type \'property\'>"
+  }
+  member {
+    name: "ram_budget"
+    mtype: "<type \'property\'>"
+  }
+  member_method {
+    name: "__init__"
+    argspec: "args=[\'self\'], varargs=None, keywords=None, defaults=None"
+  }
+}