--- a/tensorflow/tools/api/golden/v1/tensorflow.data.-options.pbtxt
+++ b/tensorflow/tools/api/golden/v1/tensorflow.data.-options.pbtxt
@@ -4,10 +4,6 @@ tf_class {
   is_instance: "<class \'tensorflow.python.data.util.options.OptionsBase\'>"
   is_instance: "<type \'object\'>"
   member {
-    name: "autotune"
-    mtype: "<type \'property\'>"
-  }
-  member {
     name: "experimental_deterministic"
     mtype: "<type \'property\'>"
   }