--- a/tensorflow/tools/api/golden/v1/tensorflow.data.experimental.-optimization-options.pbtxt
+++ b/tensorflow/tools/api/golden/v1/tensorflow.data.experimental.-optimization-options.pbtxt
@@ -8,22 +8,6 @@ tf_class {
     mtype: "<type \'property\'>"
   }
   member {
-    name: "autotune"
-    mtype: "<type \'property\'>"
-  }
-  member {
-    name: "autotune_buffers"
-    mtype: "<type \'property\'>"
-  }
-  member {
-    name: "autotune_cpu_budget"
-    mtype: "<type \'property\'>"
-  }
-  member {
-    name: "autotune_ram_budget"
-    mtype: "<type \'property\'>"
-  }
-  member {
     name: "filter_fusion"
     mtype: "<type \'property\'>"
   }