This PR adds the `tf.data.AutotuneOptions` class, which is a split of options from `tf.data.experimental.OptimizationOptions`.

- [x] add new proto message for `AutotuneOptions` and extend the `Options` message.
- [x] add `tf.data.AutotuneOptions` class.
- [x] add `autotune` attribute to `Options`.
- [x] regenerate golden APIs.
- [x] handle backward-comaptibility with `options.experimental_optimization` attributes.
- [x] add/modify `options_test` w.r.t the new `autotune` attribute.
- [x] switch to `autotune_options` in:
    - [x] `tensorflow/core/kernels/data/finalize_dataset_op.cc`
    - [x] `tensorflow/core/data/dataset_utils.cc`
    - [x] `tensorflow/core/data/root_dataset.cc `
- [x] update RELEASE.md.

TEST LOG
```
IINFO: Elapsed time: 42.805s, Critical Path: 41.39s
INFO: 354 processes: 130 internal, 224 local.
INFO: Build completed successfully, 354 total actions
//tensorflow/python/data/kernel_tests:options_test                       PASSED in 3.6s

INFO: Build completed successfully, 354 total actions

---------------

INFO: Found 13 test targets...
INFO: Elapsed time: 23.775s, Critical Path: 22.98s
INFO: 44 processes: 31 internal, 13 local.
INFO: Build completed successfully, 44 total actions
//tensorflow/python/data/experimental/kernel_tests/optimization:map_parallelization_test (cached) PASSED in 3.8s
//tensorflow/python/data/experimental/kernel_tests/optimization:autotune_buffer_sizes_test PASSED in 5.0s
//tensorflow/python/data/experimental/kernel_tests/optimization:filter_fusion_test PASSED in 13.6s
//tensorflow/python/data/experimental/kernel_tests/optimization:grappler_test PASSED in 5.6s
//tensorflow/python/data/experimental/kernel_tests/optimization:grappler_test_gpu PASSED in 5.7s
//tensorflow/python/data/experimental/kernel_tests/optimization:hoist_random_uniform_test PASSED in 6.1s
//tensorflow/python/data/experimental/kernel_tests/optimization:map_and_batch_fusion_test PASSED in 5.6s
//tensorflow/python/data/experimental/kernel_tests/optimization:map_and_filter_fusion_test PASSED in 9.0s
//tensorflow/python/data/experimental/kernel_tests/optimization:map_fusion_test PASSED in 9.8s
//tensorflow/python/data/experimental/kernel_tests/optimization:noop_elimination_test PASSED in 6.6s
//tensorflow/python/data/experimental/kernel_tests/optimization:reorder_data_discarding_ops_test PASSED in 5.5s
//tensorflow/python/data/experimental/kernel_tests/optimization:shuffle_and_repeat_fusion_test PASSED in 5.5s
//tensorflow/python/data/experimental/kernel_tests/optimization:optimization_test PASSED in 23.0s
  Stats over 2 runs: max = 23.0s, min = 7.0s, avg = 15.0s, dev = 8.0s

```

cc: @jsimsa Let me know if I have missed something. Thanks!