@jsimsa 2021-06-16
Thank you! Would it be possible for this CL to avoid the redundancy between `experimental_optimization.autotune*` options and `autotune.*` options? Could we remove the former and add a mechanism that would reroute getting and setting of the deprecated attributes to the new attributes? (similar to my changes to ThreadingOptions)